requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-yasg>=1.20
psycopg2-binary>=2.9
python-dotenv
gunicorn
